commonw: move and expand column-key tests
